refactor express in preparation for ci
