+++ app/app.js
@@ -2,9 +2,9 @@ import express from "express";
 import bodyParser from "body-parser";
 import cors from "cors";
 import dotenv from "dotenv";
-import convertCamelToSnake from "./middleware/camelToSnake.js";
-import gamerProfile from "./routes/gamerProfile.js";
-import gamerActivity from "./routes/gamerActivity.js";
+import convertCamelToSnake from "./middleware/camel-to-snake.js";
+import gamerProfile from "./routes/gamer-profile.js";
+import gamerActivity from "./routes/gamer-activity.js";
 
 dotenv.config();
