+++ app/package.json
@@ -18,6 +18,6 @@
     "pg": "^8.12.0"
   },
   "scripts": {
-    "test": "NODE_ENV=test mocha tests/**/*.js --exit"
+    "test": "NODE_ENV=test mocha '**/*.test.js' --ignore 'node_modules/**' --exit"
   }
 }
